// web/legacy_web_app/src/lib/components/Backgrounds/types/types.ts
export type QualityLevel = "high" | "medium" | "low" | "minimal";